// applets/radarr/api/model/movieHistoryEventType.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Radarr
 * Radarr API docs
 * OpenAPI spec version: 3.0.0
 */

export type MovieHistoryEventType =
  (typeof MovieHistoryEventType)[keyof typeof MovieHistoryEventType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MovieHistoryEventType = {
  unknown: "unknown",
  grabbed: "grabbed",
  downloadFolderImported: "downloadFolderImported",
  downloadFailed: "downloadFailed",
  movieFileDeleted: "movieFileDeleted",
  movieFolderImported: "movieFolderImported",
  movieFileRenamed: "movieFileRenamed",
  downloadIgnored: "downloadIgnored",
} as const;